Only use a feature if it is enabled
Change the behavior to check the target was defined and the feature was
enabled before setting it to be used/built. This is useful when the
project is built as part of other projects that may have found a package
but do not necessarily want that feature to be built by this project.