Fix the undefined symbol linking issue for the logger used in REALM_ASSERT (#369)
REALM_ASSERT uses the `log_runtime`, which is not visible to
applications. The compiler_support.h could be included by applications,
so we will see the undefined error. See
https://github.com/StanfordLegion/realm/issues/363#issuecomment-3578922317

In this PR, I moved the REALM_ASSERT into an own header, such that we do
not have to do forward declaration due to the circular include. Then we
use a dedicated logger (`log_assert`) for REALM_ASSERT, and expose it
via the librealm.so.