Fix path normalization for pkgconfig to account for absolute paths (#375)
Fixes https://github.com/StanfordLegion/realm/issues/373

pkgconfig paths are tricky because (a) the end user may relocate the
installation, and (b) when invoking CMake, the user may potentially
specify absolute paths for both `CMAKE_INSTALL_PREFIX` and
`CMAKE_INSTALL_LIBDIR`, leading to situations where one is not contained
in the other.

This patch does a couple of things:

 * If libdir (and friends) are relative, install as `${prefix}/libdir`
* If libdir is absolute but contained within the installation prefix,
make it relative and then do the above
* If libdir is absolute and NOT contained within the installation
prefix, just keep it as it. The installation will not be relocatable but
I'm not sure what else can be done in this scenario

I think this covers all the possible cases.

Tested by hand with several different configuration options, such as
setting libdir to something entirely unrelated (e.g., `/a/b/c`).